Add glitch to hypospray and make glitch common
